requirements.txt
meshtastic>=2.3.0
pypubsub>=4.0.3
llama-cpp-python>=0.2.0  # LLM chatbot support (optional)